mips: Fix invalid /reserve/ dts node
BE U-boot passes the kernel parameters, initrd addresses as well
as reserved regions by rewriting the kernel dts file. The problem is
hidden in the way the U-boot and kernel treat address space. While
kernel does have a real sight on the physical and virtual mapping,
U-boot doesn't. It is working within KSEG0 region, but thinks as it
would be physical memory region. As a consequence U-boot writes
invalid addresses to reserved dts-node. We fixed it by converting
the changed values in the fdt-validation callback to be real physical
addresses understandable by kernel.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>