mips: cache: Add secondary cache init method
In addition to asm-code, which resets the L2 cache, Baikal
Electronics added a set of functions which provides to invalidate,
writeback, prefetch, probe and initialize the secondary caches. The
code uses __fls/__ffs functions, so we need to have them defined in
the code as well.

Original code is taken from BaikalSDK 4.13

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>