baikal: Add U-boot rsa keys
It is possible to have FIT images signed by a pregenerated rsa keys.
In this case if U-boot has been built with CONFIG_FIT_SIGNATURE option
enabled, it shall check whether the FIT image is properly signed by
a public key supplied in the U-boot dtb. By this we can protect
U-boot from executing invalid Linux image.

This code is BE doing and has nothing to do with vanilly U-boot. We
keep it around for compatibility with BaikalSDK sources.

Original code is taken from BaikalSDK 4.14

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>