uboot: Add openssl path
Since U-boot mkimage utility depends on openssl for instance for
FIT-images signature check, BE built and added the precompiled library
to SDK. U-boot build system needs to know the path to it. But we won't
set the library being a part of it, since it might break Yocto build.
If you need it, jut copy and paste it from the original SDK. Otherwise
to have the system one around is enough.

Original code is taken from BaikalSDK 4.14

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>