mips: cache: Flush L2 cache at the generic cache flush methods
We need to also flush the L2 cache when generic code demands to flush
any cache. Standard code doesn't support it, since it didn't enable
the secondary cache in the first place. If we didn't added the L2-flush
here we would have got a lot of troubles in drivers.

Original code is taken from BaikalSDK 4.14

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>