baikal: Add nvram interface
Baikal-T based boards produced by Baikal Electronics are usually
equipped with a AB-RTCMC-32.768Khz-EOZ9-S3, which is supposed to be
used as RTC as well as for system NVRAM (8 byte SRAM) and as EEPROM
for serial number. But in fact It isn't really used much (oem chip is
utilised instead), while the interface still exists in the U-boot code.

Original code is taken from BaikalSDK 4.13

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>