uboot: Disable store-merging compiler optimization
This optimization is enabled by default starting from gcc 7.0.
It does work for most of the code, but for MIPS it might cause
sudden exceptions. The reason of them being rised is in unaligned
store instructions executed. Particularly it concerns static buffers
casted to an unpacked structure. Concequenced access to it fields
may cause the exception if the buffer base address turned to be
unaligned with respect to the structure fields. So to speak
we need to disable the optimization to make U-boot working
when built with newer compiler versions.

Original code is taken from BaikalSDK 4.14

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>