sata: dwc: Fix dcache invalidation for NULL buf
Really the D-cache invalidation methods doesn't permit to invalidate
NULL-based and zero-lengthed buffers. It leads to the tlb_handler whining:
tlb_handler: physical address is not found for
It might be MIPS-specific though, since noboby had had such a problem before
so to send over any fixup patch for it.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>