baikal: Discard MODULE board support
Original Module board is based on enginnering Baikal-T SoC. It was
supplied to BE in the first place, but then support was entrusted
to T-platforms back. Since TP renamed the project to SFBT1 and we are
going to have a bit more generic interface for all TP boards, lets
discard Module target support from U-boot for now.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>