baikal: Add T-platforms SFBT1 board support
SFBT1 is actually mezzonine module. But since it's fully functional device,
which only needs a motherboard with mostly driverless peripherals, we wanna
have a separate build target for it. The system is based on enginnering
version of Baikal-T (see the device datasheet for details if you got NDA
of course).

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>