baikal: video: Add SM750/video console initialization code
We don't use common CFB-console interface. Instead we just declared
private vputc and vputs method, which are used in the console
subsystem to print chars and strings to the sm750 framebuffer. We also
added the T-platforms logo printout. As the result to make magic
happening the macros CONFIG_VIDEO, CONFIG_VIDEO_SM750 and
CONFIG_SM750_CONSOLE shoule be defined in the board config file.

Signed-off-by: Alexander Matveev <alexander.matveev@t-platforms.ru>
Signed-off-by: Serge Semin <fancer.lancer@gmail.com>