baikal: Add T-platforms MITX board initialization code
The origina code has been created by Konstantin for the mITX-project.
It was based on the old U-boot code and got strong integration into
the Baikal Electronics BFK board sources. This commit is an attempt
to create a separate mITX board subsystem avoiding strong links with
the other boards sources. What comes with this patch:
- New DDR init firmware blob
- board_early_init_r, board_late_init and board_eth_init initialization,
- mITX boot configuration subblock conerning board-specific init of BMC,
  power button and so on,
- mITX-specific service command: CPU temperature, CPU warmup, board info
  loaded from FRU, memory info, MMC speed, shred,
- board-specific U-boot configurations header.
- Add delay in Realtek PHY init

Signed-off-by: Konstantin Kirik <konstantin.kirik@t-platforms.ru>
Signed-off-by: Serge Semin <fancer.lancer@gmail.com>