gpio: pca953x: Add i2c bus number config
Since most of our pca9535 expanders installed on bus 1 (not 0), we
need to have corresponding configuration. I added a new bus-field
to the width structure so to able to set the bus identifier stright
from boards configuration file. It won't affect the other boards,
since the field is appended to the others and will get defaulted
to zero if not specific in initializer.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>