gpio: Add pcf857x chip driver
This driver is partly based on pca953x code, but since the devices
are very different, the similarity is only in the file structure.
This device got quasi-bi-directional io ports. It means, that
setting one to the port is making it being weakly pulled up, so
it might be treated as input mode. If ports bit is set to zero,
the output will be strongly pulled down. So the driver prvides
just two operations: get and set value.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>