baikal: Add T-platforms T-client board initialization code
This device is in fact the carrier board for mezzanine modules
SFBT1/MSBT2. In addition to the basic modules functions, we
enable PCIe and video driver here since the device got SM750
on board. PIO-expanders placed on the carrier device is also
avaialble for customization. board_early_init_r and board_late_init
methods are also defined empty for now. DW Ethernet controller is
reset for MSBT2-version only and initialised in board_eth_init.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>