baikal: Add T-platforms CNC board initialization code
This device series is the set of carrier boards for mezzanine modules
SFBT1/MSBT2. In addition to the basic modules functions, we added
the support of pcf8574 installed on each board and serving as
a shred with board id and revision. We are using this info to determine
the FIT configuration so the kernel would be loaded with proper
dtb. The device detection is added to board_late_init stage.
board_early_init_r is still left empty. DW Ethernet controller is reset
for MSBT2-version only and initialised in board_eth_init.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>