net: designware: De-assert GMAC GP out for any board
This is kind of a traditional issue of the Baikal-T(1) based boards.
Each of them got ethernet PHY reset pin connected to the GMAC GP out
pin. It means when the GMAC gets to be reset, PHY also turns into the
reset state. In order to make things working, we need to de-assert
the GP out pin each time the GMAC is reset. Since the issue persists
nearly on all Baikal-T devices, lets make the fix available when
CONFIG_TARGET_BAIKAL_MIPS is declared.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>