autoboot: Update menukey on autoboot-keyed
Autoboot-keyed mode provides a functionality to create a magic string,
which when is netered can abort the autoboot and expose the U-boot
console. But autoboot process doesn't provide a way to enter the bootmenu
right after the magic string is entered. It is only provided by normal
non-keyed boot. We fixed it just by using the first matched key symbol
to be assigned to the menukey. In case if it also matches CONFIG_MENUKEY,
then command from "menucmd" evironment variable is executed.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>