phy: mscc: Call basic init script on startup
U-boot really likes to reset ethernet MAC and as a consequence it PHY
as well. If we call ping or perform any network activity, all the
interfaces get to be reset. But since usually only one of them is used
the other one is left unconfigured. In this case lets at least perform
vsc8531/vsc8541 init-scripts method so to have LEDs in pre-defined state.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>