uboot: Add default bootmenu entry functionality
It is handy to have a default bootmenu entry to be selected
instead of entry 0 as it was before. This commit provides such
a functionality. User just need to set "bootmenu_default" variable
with a number of corresponding entry and one will be selected
when the timer gets to zero.

Original code is taken from BaikalSDK 4.14

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>