ext4: Fix 64bit flag incompatibility
The latest and greatest e2fsprogs (>1.43) has turned on 64bit feature
for EXT4 filesystems. The U-Boot will fail to understand this kind of
filesystem and until this commit will bail out with the aforementioned
messages.

The fix would be either to disable 64bit flag or to have this commit
applied.

Signed-off-by: Ilya Novikov <Ilya.Novikov@t-platforms.ru>