baikal: bc3bt1: Add keyed menu entry
Since the system is going to be utilized in a complicated environment
there might be noise on the serial lines. In this case the boot process
may be interrupted by entering to the interactive boot menu. In order
to minimize the possibility of this lets enable the keyed menu feature,
so in order to get into the interactive menu the user shall enter
a symbol 'm' before the timer expired.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>