baikal: cnc: Set fdt_file_name when board detected
CNC firmware is created with multi-config FIT image support. In particular
the board type is determined on U-boot start so the corresponding
FIT-config would be selected. But in case net-boot FIT image isn't used.
Instead the component bootload proceeds. In this case we also need
to upgrade the fdt_file_name env variable with proper dtb file name.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>