baikal: Add custom setting of DDR clocks frequency
In case if it necessary to drop a system power concumption simultaniusly
decreasing device heating, a user can decrease the memory bus frequency
by new CONFIG_DDR_CUSTOM_CLK config. It can vary from 1066 to 1600, but
eventually will get rounded to either 1066, 1333 or 1600. The functionality
as developed on top the SPD-read and DDR initialization code. It just
recalculates the tCLK (real clocks period) in accordance with the value
of config.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>
Fixes: #14