baikal: Append dtb to U-boot binary instead of embedding it
Even though CONFIG_OF_EMBED is properly working well now. Lets
use an appended dtb file instead of embedding it. It is handy for
multiple reason, one of the is due to Yocto build process, which
appends the blob to the u-boot-binary anyway. This feature is
enabled by config CONFIG_OF_SEPARATE.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>
Fixes: #15