Baikal: Improve PCIe initialization
Some PCI devices require the Host Controller to be ready for link
training in a short time after device reset. If the HC appears not
ready this may result in permanent host failure till the next hard
reset.

It seems easier to address the problem in u-boot rather than kernel.
We do it in a way compatible to linux, so reconfiguring the PCIe
bus is not needed.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>