baikal: bfk: Create new config for cmd_service
Baikal-T BFK board-specific got its own cmd_serice module, and it's
enabled by CONFIG_CMD_SERVICE config. In case if we get to build the
U-boot image for BFK, we'll come up with an error because there is
a T-platforms board-specific cmd_service activated with the same
config. So in order to fix it lets rename the BFK cmd_service config to
CONFIG_BFK_CMD_SERVICE.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>