baikal: Enable FRU for MSBT2/MRBT1-based boards
First of all we need to make sure the TP_FRU config is enabled
for MSBT2/MRBT1 mezzanine cards and for the boards equipped with
ones. It is done just by defaulting TP_FRU config for MSBT2 and
MRBT1 and their carrier boards.

Secondly we need to call the FRU open-parse function from the
board-specific code. It should be done unconditionally for MSBT2
and MRBT1 (also fo BC3BT1) specific code, while init methods of
CNC, Axitech1 and T-client carrier boards should init FRU only if it
is enabled. It is done so since they can be equipped with SFBT1 card,
which doesn't have FRU on it.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>