mips: baikal: Enable CONFIG_CMD_SAVEENV for mITX
This config has been commented out since the times when saveenv
command was broken due to unsupported DW SPI boot interface. Recently
BE fixed the problem with the interface, so the SPI interface is now
cappable to interact with SPI-flash connected to it the write operations.
So in order to be able to perform the U-boot environment save operation
we need to enable the command for mITX devices as we already did for the
rest of TP boards.

Closes #6

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>