mips: baikal: Create board-specific PCI/USB reset method
In order to make the Baikal-architecture code more generic lets
create a common board_pci_reset() method, which in fact is already
defined in nearly all TP boards-specific sources. We also need to
take into account a case when CONFIG_PCIE_DW config is enabled,
which already have the board_pci_reset() method invocation. For
this case we create a weakly defined default method.

Similar method is created for USB, though it is used for mITX and
partly for bn1bt1 projects.

Fixes #22

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>
Signed-off-by: Serge Semin <fancer.lancer@gmail.com>