USB: xHCI: Fix physical addresses passed to controller
Platforms which do not provide 1-to-1 memory mapping in u-boot
require conversion of (virtual) addresses passed to controller.

To achieve that, we define __pa() and __va() macros based on
platform dependent virt_to_phys()/phys_to_virt(). On platforms
with 1-to-1 mapping these macros are optimized out by compiler.

The __pa() macro is used wherever physical address is passed to the
controller.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>