MIPS: Baikal: Do not clobber t0 register on exception
Exception and interrupt entry points use t0 to jump to the
handler. This clobbers the register and may lead to unexpected
behaviour of the code that caused exception.
Replace it with t9 register. This does not resolve all possible
problems but seems sufficient if exceptions are caused by
memtest code accessing unmapped memory.

Signed-off-by: Vadim V. Vlasov <vadim.vlasov@t-platforms.ru>