mips: baikal: Reuse lowmem size declared in fdt
Due to a stupid solution of our older kernel developers, the low memory
region was reduced by 32MB. It was done so to expel the range from the
system memory so to make use of it for a video frame buffer. Modern
kernels don't need to have this approach utilized since they have a full
support of the reserved-memory nodes. But since the bootloader must
be able to load both older and new kernels, that broken solution need to
be also supported. In order to do this we read the existing size of the
low memory from the passed fdt and then just resave it back to the dt-blob.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>