mips: baikal: Reduce MAX highmem size to be 0xDFFF0000
Old Linux kernels don't support dts with XPA-extended memory cells.
They also fails to have a high memory of maximum size 0xE0000000
due to a bug in the MIPS architecture code. In order to bypass it
we have to reduce maximum high memory size to be reduced with 64KB.
By doing so we prevent the memory range overflow and let the kernel
to decrease the upper limit with a size of the memory page.

Signed-off-by: Serge Semin <fancer.lancer@gmail.com>